Move parameter configuration into zeitgeist module
